Extract ImageOrVideo component
