+++ src/Widgets/CardWidget/CardWidgetComponent.tsx
@@ -8,10 +8,10 @@ import {
   Link,
   LinkTag,
   isInPlaceEditingActive,
-  Widget,
 } from 'scrivito'
 import { CardWidget } from './CardWidgetClass'
 import { alternativeTextForObj } from '../../utils/alternativeTextForObj'
+import { ImageOrVideo } from '../../Components/ImageOrVideo'
 
 provideComponent(CardWidget, ({ widget }) => {
   const cardBodyClassNames: string[] = ['card-body']
@@ -91,44 +91,3 @@ const LinkOrNotTag = connect(
     )
   },
 )
-
-const ImageOrVideo = connect(function ImageOrVideo<T extends string>({
-  attribute,
-  className,
-  widget,
-}: {
-  attribute: T
-  className?: string
-  widget: Widget<{ [K in T]: 'reference' }>
-}) {
-  const background = widget.get(attribute)
-  if (!background) return null
-
-  const classNames = ['img-background']
-  if (className) classNames.push(className)
-
-  if (background.contentType().startsWith('video/')) {
-    return (
-      <video
-        autoPlay
-        className={classNames.join(' ')}
-        key={background.contentUrl()}
-        loop
-        muted
-        playsInline
-        src={background.contentUrl()}
-      />
-    )
-  }
-
-  return (
-    <InPlaceEditingOff>
-      <ImageTag
-        content={widget}
-        attribute={attribute}
-        className={classNames.join(' ')}
-        alt=""
-      />
-    </InPlaceEditingOff>
-  )
-})
